1.0
Extras added to the forms, such as the "other" gender extra years of
your birthday up to 2018 and some upcoming updates page link on the
sticky sidebar, also moved a small thing in CSS, need to work on design
soon to make it so it works. and then all the elements can make a good
looking website.